requirements.txt
streamlit
pandas
numpy
pydantic
scikit-learn
google-genai 